Formula arreglada

+++ foodrat.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Servidor: 127.0.0.1
--- Tiempo de generación: 05-10-2023 a las 21:11:42
+-- Tiempo de generación: 08-03-2024 a las 10:47:57
 -- Versión del servidor: 10.4.25-MariaDB
 -- Versión de PHP: 8.1.10
 
@@ -30,12 +30,12 @@ SET time_zone = "+00:00";
 CREATE TABLE `diadieta` (
   `idrat` int(2) NOT NULL,
   `fecha` datetime NOT NULL,
-  `peso` int(3) NOT NULL,
-  `sobras` int(3) NOT NULL DEFAULT 0,
-  `dieta` int(2) DEFAULT NULL,
-  `diferencia` int(4) DEFAULT NULL
+  `peso` float NOT NULL,
+  `sobras` float NOT NULL DEFAULT 0,
+  `dieta` float DEFAULT NULL,
+  `diferencia` float DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
--- --------------------------------------------------------
+
 --
 -- Estructura de tabla para la tabla `rat`
 --
@@ -44,9 +44,8 @@ CREATE TABLE `rat` (
   `idrat` int(2) NOT NULL,
   `name` varchar(5) NOT NULL,
   `fase` int(1) NOT NULL DEFAULT 1,
-  `pesoestable` int(4) DEFAULT NULL
+  `pesoestable` float DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
-
 --
 -- Indices de la tabla `rat`
 --
